Addding exp_safe_html transform
