+++ src/experimental/safe_html_transform/profiles/default/metadata.xml
@@ -3,5 +3,7 @@
   <version>1000</version>
   <dependencies>
   	<dependency>profile-plone.app.registry:default</dependency>
+  	<dependency>profile-Products.MimetypesRegistry:MimetypesRegistry</dependency>
+	<dependency>profile-Products.PortalTransforms:PortalTransforms</dependency>
   </dependencies>
 </metadata>
