+++ src/experimental/safe_html_transform/setuphandlers.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
+import pkgutil
 import logging
-from zope.component import getUtility
-from Products.PortalTransforms.interfaces import IPortalTransformsTool
 from Products.CMFCore.utils import getToolByName
+import experimental.safe_html_transform.transforms
 
 
 def isNotCurrentProfile(context):
@@ -14,23 +14,35 @@ def post_install(context):
     if isNotCurrentProfile(context):
         return
     # Do something during the installation of this package
-    unregister_transform(context, 'safe_html')
+    uninstallOldTransform(context)
+    installNewTransform(context)
 
 
-def unregister_transform(context, transform):
-    transform_tool = getUtility(IPortalTransformsTool)
-    if hasattr(transform_tool, transform):
-        transform_tool.unregisterTransform(transform)
+def availableTransforms():
+    try:
+        return [x[1] for x
+                in pkgutil.iter_modules(experimental.safe_html_transform.transforms.__path__)
+                if x[1] != 'mimetypes']
+    except ImportError:
+        return []
 
 
-def installTransform(context, logger=None):
+def uninstallOldTransform(context, logger=None):
     if logger is None:
         logger = logging.getLogger('experimental.safe_html_transform')
-    if hasattr(context, 'readDataFile'):
-        if isNotCurrentProfile(context):
-            return
-    else:
-        pass
+    pt = getToolByName(context, 'portal_transforms')
+    # get name of safe_html transform from our add-on.
+    transforms = 'safe_html'
+    # check for our transform name in the portal_tranform
+    if transforms in pt:
+        pt.unregisterTransform(transforms)
+        logger.info('Unregistered %s' % transforms)
+
+
+def installNewTransform(context, logger=None):
+    if logger is None:
+        logger = logging.getLogger('experimental.safe_html_transform')
     transforms = getToolByName(context, 'portal_transforms')
-    transforms.unregisterTransform('safe_html')
-    transforms.manage_addTransform('safe_html', 'experimental.safe_html_transform.%s' % 'safe_html')
+    for tform in availableTransforms():
+        transforms.manage_addTransform(tform, 'experimental.safe_html_transform.transforms.%s' % tform)
+        logger.info('Registered %s' % tform)
